fixup/clarify slipstream's behaviour + fix tests
The slipstream() method was used inconsistently between supported games.
The original implementation (with doom) would take a path and a file, and
put the file into a .deb package at the path supplied - /including the
filename/, e.g.

  slipstream $deb foo/bar/baz/doom2.wad doom2.wad
    => file ./foo/bar/baz/doom2.wad
  slipstream $deb foo/bar/baz/doom2.wad DOOM2.WAD
    => file ./foo/bar/baz/doom2.wad
  slipstream $deb foo/bar/baz doom2.wad
    => file ./foo/bar/baz

Later, slipstream was adjusted to support multiple file arguments, which
meant the path was interpreted as a parent directory, rather than a final
path name for one file.  The quake games needed this behaviour, and
slipstream was hacked to behave each way depending on the presence of a
'/' suffix.

This patch adjusts slipstream to always behave wiht the supplied path
being a directory, rather than a final pathname for the file. The
/-suffix behaviour switch is no longer needed. Adjust all callers as
necessary.

This introduces a bug for the doom packages: the filename of the wad
file is not canonicalized to doom.wad (or tnt.wad or whatever), so some
extra code is needed to resolve that.

Finally, fix the slipstream test in the testsuite. Slipstream now also
creates the path if it does not already exist.