quake2: avoid an extra copy by using files instead of a skeleton .deb
The data files are pretty large. We can avoid one copy if instead of
assembling the data files in a staging area, unpacking a skeleton .deb,
copying the data files into the hierarchy produced by that skeleton
and repacking it, we just assemble them into a hierarchy matching
what we want in the .deb, inject the control files and pack it up.