quake: rework to support downloading shareware 1.06
Now that we can auto-download quake106.zip from an idstuff mirror,
we don't really need to support installing shareware version 1.01
any more (no point), so we can be more specific about what we require.

Similarly, because the 1.06 shareware pak0.pak is actually the
same file as the 1.01 shareware pak0.pak, we can use quake106.zip
as (effectively) a patch for the full version.