wolf3d: rename "anthology" to "wl6a", add more annotations
"anthology" seems a confusing abbreviation because Wolf3D is available
in (at least) the id Anthology, which is the wl6 flavour, and the
3D Realms Anthology, which is the wl6a flavour. Let's call it wl6a
because that's how Wolf4SDL distinguishes this version.

skitt's file listing included a file "config.wl6" which is not in
any other edition. I'm not adding it to the files that can be
packaged right now, because it isn't clear whether it's needed.