Make GameDataPackage.type read-only, populated from demo_for/expansion_for
Use demo_for and expansion_for directly in most cases.

This lets us improve the special handling of demo packages in one
corner case: previously we would not build a demo package if we were
building any full game, now we will not build a demo package if we
are building the full game directly corresponding to that demo.