rework engine definition, avoid need for duplicated data
The 'debian: engine:' tag is moved to top-level 'engine:'
tag that can be overiden at package level if needed.

This let avoid the need to specify the the same engine
for a game and it's demo.

The doom_common.py plugin is reworked to derive the engine
program binary from the engine package name;
again to avoid duplicated data.

The current default doom-engine is now set in an unique location;
and it isn't needed to explicitely set anymore in all the .yaml
files using the doom_common.py plugin.