hacx: distributable, but still non-free
"Hacx can now be distributed freely by any site that wishes to host it.
Also, any site that wishes to create a distribution area for files, mods,
bug fixes, etc., please, feel free to do so.

Hacx is now a royalty free item. This means that Hacx may NOT be used to
generate income of any kind nor can it be used to promote any activity that
would generate income."