make_template: make better use of groups
Each file found is now placed in one of several groups:
"probably required", "probably optional", "probably documentation",
"probably licenses", "probably unwanted" and "archives".

As a step towards sharing code between make-template and
the rest of g-d-p, we use the same WantedFile objects for groups
as we do in the main g-d-p codebase.