PackagingTask: distinguish docdir with and without destdir prepended
In the symlink-creation code, we were actually getting it wrong
for packages like quake3 that use $docdir, and trying to create a
symlink from /tmp/.../usr/share/doc/... to /tmp/.../usr/share/doc/...;
but because os.path.join(destdir, "/tmp/...") == "/tmp/...",
and because we have code to make the symlinks as relative as possible,
it worked anyway.

Similarly, distinguish licensedir with and without destdir prepended.