run_command_line: build all packages if argument matches game name
Normally, a game "foo" builds packages that include "foo-data" but
not "foo", so it's OK to have the shorthand that "g-d-p foo-data"
builds exactly that package and no more. However, this breaks down
if we want a game to build a package whose name is the same as the
game itself.

We avoided this for Unreal entirely by mistake, because we don't
build an unreal.deb (because unreal-classic and unreal-gold are
co-installable), but for Unreal Tournament we want to make the
"main package" match the game name.