d/control: consistently use Arabic numerals and "and" to denote sequels
Some of these are correctly titled with Roman numerals like
Quake III Arena, and some are correctly Arabic numerals like
Quake 4, but the list is considerably easier to keep organised if we
pick one.

Use "and" to separate lists, so that "&" is unambigously part of the
title (for games like Heroes of Might & Magic).