Avoid two-pass processing of groups files
Since commit 2b5a38e0 this has been crashing with io.UnsupportedOperation
"underlying stream is not seekable" for some people and in some
situations, but not consistently. Duplicating the list of group names
at the beginning so that we can process the file linearly is not very
expensive, and should be mitigated by the compression anyway.