declare compliance with Standards-Version 3.9.7 **
https://www.debian.org/doc/debian-policy/upgrading-checklist.html

At first
"recommend to ship additional documentation for package pkg
in a separate package pkg-doc and install it into /usr/share/doc/pkg"
seemed relevant for generated larry-doc, but as this documentation
booklet is shared accross 5 different games; this doesn't apply.

For other games, the documentation:
- is most of the time dwarfed by the game assets
- is already downloaded in the same archive anyway
so we can as well just keep it as-is too.