packaging: allow systems to override pkgdocdir, pkglicensedir
I'm hoping to add support for generating Flatpak extensions containing
the non-distributable game data, along the same lines as
<https://cgit.freedesktop.org/xdg-app/freedesktop-sdk-images/tree/make-nvidia-image.sh>
for nVidia drivers. Each extension only contains one directory,
and I'm planning to use $install_to for that, so we'll need to
redirect $pkgdocdir and $pkglicensedir to be below $install_to.