Partially revert "remove some src:quake-imported weirdness"
These dependencies on the Makefile were intentional. Without a
dependency, a change to the Makefile recipe will not trigger an
update to the generated file, which is bad if some significant
feature of the generated file (in this case the colour shifting
that is applied to it) is written in the Makefile.

This partially reverts commit 7e8f27c9f1c4e65e27919ca4989df1dac1f722ac.